+++ fastai/imports/core.py
@@ -3,7 +3,7 @@ import math, matplotlib.pyplot as plt, numpy as np, pandas as pd, random
 import scipy.stats, scipy.special
 import abc, collections, hashlib, itertools, json, operator, pathlib
 import mimetypes, inspect, typing, functools
-import html, re, spacy, requests, tarfile
+import html, re, spacy, requests, tarfile, numbers
 
 from abc import abstractmethod, abstractproperty
 from collections import abc,  Counter, defaultdict, Iterable, namedtuple, OrderedDict