+++ fastai/vision/models/unet.py
@@ -14,7 +14,7 @@ def _get_sfs_idxs(sizes:Sizes) -> List[int]:
 class UnetBlock(nn.Module):
     "A quasi-UNet block, using `PixelShuffle_ICNR upsampling`."
     def __init__(self, up_in_c:int, x_in_c:int, hook:Hook, final_div:bool=True, blur:bool=False, leaky:float=None,
-                 self_attention:bool=False,**kwargs):
+                 self_attention:bool=False, **kwargs):
         super().__init__()
         self.hook = hook
         self.shuf = PixelShuffle_ICNR(up_in_c, up_in_c//2, blur=blur, leaky=leaky, **kwargs)
@@ -38,7 +38,8 @@ class UnetBlock(nn.Module):
 class DynamicUnet(SequentialEx):
     "Create a U-Net from a given architecture."
     def __init__(self, encoder:nn.Module, n_classes:int, blur:bool=False, blur_final=True, self_attention:bool=False,
-                 sigmoid:bool=False, last_cross:bool=True, bottle:bool=False, **kwargs):
+                 y_range:Optional[Tuple[float,float]]=None,
+                 last_cross:bool=True, bottle:bool=False, **kwargs):
         imsize = (256,256)
         sfs_szs = model_sizes(encoder, size=imsize)
         sfs_idxs = list(reversed(_get_sfs_idxs(sfs_szs)))
@@ -56,8 +57,8 @@ class DynamicUnet(SequentialEx):
             up_in_c, x_in_c = int(x.shape[1]), int(sfs_szs[idx][1])
             do_blur = blur and (not_final or blur_final)
             sa = self_attention and (i==len(sfs_idxs)-3)
-            unet_block = UnetBlock(up_in_c, x_in_c, self.sfs[i],
-                                   final_div=not_final, blur=blur, self_attention=sa, **kwargs).eval()
+            unet_block = UnetBlock(up_in_c, x_in_c, self.sfs[i], final_div=not_final, blur=blur, self_attention=sa,
+                                   **kwargs).eval()
             layers.append(unet_block)
             x = unet_block(x)
 
@@ -68,7 +69,7 @@ class DynamicUnet(SequentialEx):
             ni += 3
             layers.append(res_block(ni, bottle=bottle, **kwargs))
         layers += [conv_layer(ni, n_classes, ks=1, use_activ=False, **kwargs)]
-        if sigmoid: layers.append(nn.Sigmoid())
+        if y_range is not None: layers.append(SigmoidRange(*y_range))
         super().__init__(*layers)
 
     def __del__(self):